Revert "Added pytest-dependency for microscope fixture and microexpansion_coords tests"
This reverts commit 4dbcc322cfff9eae8884de32da61b763457d35b4.
